+++ modules/bot/views/publics/top/index.php
@@ -1,0 +1,4 @@
+Top users list: <br/>
+<?php foreach ($users as $user) {?>
+<?= $user['username'] ?>: <?= $user['rating'] ?><br/>
+<?php } ?>
