+++ modules/bot/views/publics/top/vote.php
